+++ src/br/com/gamestore/persistences/VendaManager.java
@@ -7,14 +7,14 @@ import br.com.gamestore.entidades.Venda;
 
 public class VendaManager {
 	
-	private static TreeMap<Integer, Venda> venda = new TreeMap<Integer, Venda>();
+	private static TreeMap<Double, Venda> venda = new TreeMap<Double, Venda>();
 	private static int cod = 0;
 
-	public static TreeMap<Integer, Venda> getVenda() {
+	public static TreeMap<Double, Venda> getVenda() {
 		return venda;
 	}
 
-	public static void setVenda(TreeMap<Integer, Venda> venda) {
+	public static void setVenda(TreeMap<Double, Venda> venda) {
 		VendaManager.venda = venda;
 	}
